pyproject.toml
[tool.pytest.ini_options]
testpaths = ["tests"]
[project]
name = "ask-hr-local"
version = "0.1.0"
description = "Local RAG: Ollama + Chroma + FastAPI + Streamlit over HR corpus"
requires-python = ">=3.11"
# If you don't have a README yet, omit the next line or add a stub README.md
# readme = "README.md"

dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.30.1",
  "pydantic==2.7.4",
  "python-dotenv==1.0.1",
  "httpx==0.27.0",
  "streamlit==1.36.0",
  "chromadb==0.5.4",
  "tqdm==4.66.4",
  "PyPDF2==3.0.1",
  "unstructured==0.15.7",
  "unstructured-client==0.23.5",
  "markdownify==0.11.6",
  "pyyaml==6.0.1",
  "tiktoken==0.7.0"
]
[project.optional-dependencies]
dev = [
  "ruff>=0.0.282,<0.4.0",
  "black>=23.7,<25.0",
  "mypy>=1.5,<2.0",
  "pytest>=7.4,<8.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Tell setuptools how to find your package (so -e . works)
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

